requirements/base.txt
django==3.2.5
djangorestframework==3.12.4
drf-yasg==1.20.0